pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ssd-agent"
version = "0.1.0"
description = "SSD pipeline auditável via Deep Agents + ACP"
requires-python = ">=3.11"

dependencies = [
    "agent-client-protocol>=0.1.0",
    "deepagents>=0.1.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ssd-agent = "ssd_agent.__main__:main"
ssd-agent-cli = "ssd_agent.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ssd_agent"]
